Usedcar Price Project
In this project, we collect the data first then we use that data build a ML model. You can see two excel sheet, car_dataset.xlsx is the one on which we build the model and right report and prepare presentation. While usedcar_dataset.xlsx is the modified one, in which we tried to scrape some more data from car24.com website.